refactor: use ESM
