+++ src/DevMiddlewareError.js
@@ -1,5 +1,3 @@
-'use strict';
-
 class DevMiddlewareError extends Error {}
 
-module.exports = DevMiddlewareError;
+export default DevMiddlewareError;
